Specify dtype in tests
Fails sometimes with np.float32 dtype:
the neighbours aren't always 100% identical.

+++ tests/test_correctness.py
@@ -9,10 +9,12 @@ from sklearn.neighbors import NearestNeighbors
 @pytest.mark.parametrize("n_neighbors", [1, 10, 100])
 @pytest.mark.parametrize("use_chunks_on_Y", [True, False])
 @pytest.mark.parametrize("chunk_size", [int(4096 * fac) for fac in [0.5, 1, 2]])
-def test_correctness(n, d, n_neighbors, use_chunks_on_Y, chunk_size):
+#@pytest.mark.parametrize("dtype", [np.float32, np.float64])
+def test_correctness(n, d, n_neighbors, use_chunks_on_Y,
+                     chunk_size, dtype=np.float64):
     np.random.seed(1)
-    Y = np.random.rand(int(n * d)).reshape((-1, d))
-    X = np.random.rand(int(n * d // 2)).reshape((-1, d))
+    Y = np.random.rand(int(n * d)).astype(dtype).reshape((-1, d))
+    X = np.random.rand(int(n * d // 2)).astype(dtype).reshape((-1, d))
 
     neigh = NearestNeighbors(n_neighbors=n_neighbors, algorithm='brute')
     neigh.fit(Y)
